local-docker/requirements.txt
pandas==2.2.3
xgboost==2.1.4
scikit-learn==1.6.1
matplotlib==3.10.0
seaborn==0.13.2
mlflow==2.20.2
minio==7.2.15
#psycopg==2.9.10
Boto3==1.36.23
python-dotenv==1.0.1

seaborn>=0.11.0
joblib>=1.1.0
pytest==8.3.5